+++ Runtime/Core/ChirpImpl.cs
@@ -9,14 +9,13 @@ namespace WhiteSparrow.Shared.Logging.Core
 {
     internal class ChirpImpl : IChirpReceiver, IDisposable, IChirpChannels
     {
+        private List<IChirpPlugin> m_Plugins = new List<IChirpPlugin>();
         private List<IChirpInput> m_Inputs = new List<IChirpInput>();
-        
         private List<IChirpOutput> m_Outputs = new List<IChirpOutput>();
-        private IChirpOutput[] m_OutputsCached;
 
         public ChirpImpl()
         {
-            m_DefaultChannel = CreateInternal("Default", 0);
+            m_DefaultChannel = new ChirpLogger("Default", useChannel: false);
             DefaultStyle = new ChirpStyle();
             // DefaultStyle.LogColorInfo = new Color(52f/255f, 195f/255f, 235f/255f);
             // DefaultStyle.LogColorWarning = new Color(235f/255f, 159f/255f, 52f/255f);
@@ -27,46 +26,59 @@ namespace WhiteSparrow.Shared.Logging.Core
 
         #region Components
 
-        public T AddInput<T>()
-            where T : class, IChirpInput, new()
+        public T AddPlugin<T>()
+            where T : class, IChirpPlugin, new()
         {
-            T input = new T();
-            m_Inputs.Add(input);
-            input.Initialize(this);
-
-            return input;
-        }
-        public void AddInput(IChirpInput input)
-        {
-            m_Inputs.Add(input);
-            input.Initialize(this);
-        }
-
-        public T AddOutput<T>()
-        where T: class, IChirpOutput, new()
-        {
-            T output = new T();
-            m_Outputs.Add(output);
-            m_OutputsCached = m_Outputs.ToArray();
-            output.Initialize();
-
-            return output;
-        }
-        public IChirpOutput AddOutput(IChirpOutput output)
-        {
-            m_Outputs.Add(output);
-            m_OutputsCached = m_Outputs.ToArray();
-            output.Initialize();
+            T instance =  new T();
+            AddPlugin(instance);
+            return instance;
             
-            return output;
         }
 
-        public void RemoveOutput(IChirpOutput output)
+        public void AddPlugin(IChirpPlugin instance)
         {
-            m_Outputs.Remove(output);
-            m_OutputsCached = m_Outputs.ToArray();
+            if (instance is IChirpInput input)
+            {
+                m_Inputs.Add(input);
+                input.InitializeInput(this);
+            }
+
+            if (instance is IChirpOutput output)
+            {
+                m_Outputs.Add(output);
+            }
+            
+            RegisterPlugin(instance);
         }
-        
+
+        private void RegisterPlugin(IChirpPlugin instance)
+        {
+            if (m_Plugins.Contains(instance))
+                return;
+            
+            m_Plugins.Add(instance);
+            instance.OnDisposed += OnPluginDisposed;
+        }
+
+        private void OnPluginDisposed(IChirpPlugin instance)
+        {
+            RemovePlugin(instance);
+        }
+
+        public void RemovePlugin(IChirpPlugin instance)
+        {
+            if (!m_Plugins.Contains(instance))
+                return;
+
+            instance.OnDisposed -= OnPluginDisposed;
+            m_Plugins.Remove(instance);
+
+            if (instance is IChirpInput input)
+                m_Inputs.Remove(input);
+            if (instance is IChirpOutput output)
+                m_Outputs.Remove(output);
+        }
+
         [HideInCallstack]
         public void Submit(ChirpLog log)
         {
@@ -74,8 +86,8 @@ namespace WhiteSparrow.Shared.Logging.Core
             
             if(log.Options.AddStackTrace)
                 PopulateStackTrace(log);
-            
-            foreach (var output in m_OutputsCached)
+
+            foreach (var output in m_Outputs)
             {
                 output.Ingest(log);
             }
@@ -95,16 +107,9 @@ namespace WhiteSparrow.Shared.Logging.Core
         public void Dispose()
         {
             var inputs = m_Inputs.ToArray();
-            m_Inputs.Clear();
+                m_Inputs.Clear();
 
-            foreach (var input in inputs)
-                input.Dispose();
 
-            var outputs = m_OutputsCached;
-            m_Outputs.Clear();
-            m_OutputsCached = null;
-            foreach (var output in outputs)
-                output.Dispose();
         }
 
         #endregion
@@ -114,30 +119,6 @@ namespace WhiteSparrow.Shared.Logging.Core
         private static int s_LoggerId = 1;
         private ChirpLogger m_DefaultChannel;
         ChirpLogger IChirpChannels.Default => m_DefaultChannel;
-
-        
-        public ChirpLogger Create(string name)
-        {
-            if (string.IsNullOrWhiteSpace(name))
-                throw new ArgumentNullException(nameof(name));
-            
-            return CreateInternal(name, s_LoggerId++, null);
-        }
-
-        public ChirpLogger Create(string name, Color color)
-        {
-            if (string.IsNullOrWhiteSpace(name))
-                throw new ArgumentNullException(nameof(name));
-            
-            return CreateInternal(name, s_LoggerId++, color);
-        }
-
-     
-        private ChirpLogger CreateInternal(string name, int id, Color? color = null)
-        {
-            ChirpLogger logger = color.HasValue ? new ChirpLogger(name, id, color.Value) :  new ChirpLogger(name, id);
-            return logger;
-        }
 
         #endregion
 
@@ -154,8 +135,6 @@ namespace WhiteSparrow.Shared.Logging.Core
     public interface IChirpChannels
     {
         ChirpLogger Default { get; }
-        ChirpLogger Create(string name);
-        ChirpLogger Create(string name, Color color);
         
     }