+++ Runtime/Core/IChirpPlugin.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace WhiteSparrow.Shared.Logging.Core
+{
+	public interface IChirpPlugin : IDisposable
+	{
+		event Action<IChirpPlugin> OnDisposed;
+	}
+}