+++ Runtime/Utils/LoggingStackTraceUtil.cs
@@ -35,7 +35,6 @@ namespace WhiteSparrow.Shared.Logging
 
 		public static string FormatUnityStackTrace(StackTrace stackTrace)
 		{
-			// channel = null;
 			var stringBuilder = s_StackTraceBuilder;
 			stringBuilder.Clear();
 			for (var index1 = 0; index1 < stackTrace.FrameCount; ++index1)
@@ -47,9 +46,6 @@ namespace WhiteSparrow.Shared.Logging
 					var declaringType = method.DeclaringType;
 					if (declaringType != null)
 					{
-						// if(channel == null)
-						// 	channel = LogChannel.GetForTarget(declaringType);
-						
 						var str1 = declaringType.Namespace;
 						if (!string.IsNullOrEmpty(str1))
 						{